(AWSGRID-27) Address Amazon pull request review issues
This commit addresses Amacon's pull request review issues in
lustre.template, specifically:

This needs to be split into multiple statements such that the resources can be restricted
for various actions. Specifically for dynamodb actions, giving permissions to '*' is not
desirable. Otherwise, the instance will be able to read any dynamodb table in the region
when that is not the intent.

Change-Id: I2915ba76fe891c4a930c919688de83100c494528